requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
yfinance>=0.2.50
matplotlib>=3.9
scikit-learn>=1.5
requests>=2.32
python-dotenv>=1.0
pydantic>=2.8
PyYAML>=6.0
openai>=1.43
tiktoken>=0.7
dotenv
google-generativeai>=0.8.5
protobuf>=4.25.3
google-ai-generativelanguage>=0.6.6
google-api-core>=2.19.0

